requirements.txt
jupyter
numpy
pandas
matplotlib
pillow
seaborn
scikit-learn
scikit-image
opencv-python
loguru
torch
torchvision
tqdm
einops
setuptools
scipy
hydra-core
